requirements.txt
discord.py==1.6.0
num2words==0.5.10
prsaw==0.2.0